#273-capitalize-types - Add auth support for field definitions
